requirements.txt
langchain
python-dotenv
streamlit
google-generativeai
